requirements.txt
# Core Web Framework
Flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
eventlet>=0.33.0
werkzeug>=3.0.0

# Database
# sqlite3 is built-in

# Image Processing
numpy>=1.26.0
opencv-python-headless>=4.8.0  # Headless is critical for cloud servers
Pillow>=10.0.0

# AI & Utilities
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyttsx3>=2.90  # Warning: Needs espeak on server, might fail init
fpdf>=1.7.2

# Face Recognition (Heavy - enable if server can handle it)
# face-recognition>=1.3.0
# dlib>=19.24.0

# Senku Autonomous Teaching
chromadb>=0.4.0
google-genai>=1.0.0
PyPDF2>=3.0.0
